Refactor SpecWatcher
This attempts to simplify and streamline the functionality around
keeping the running services in agreement with the specs that define
them on disk.

An important component of this is pulling the logic that determines
what should start, stop, and restart out of the SpecWatcher itself,
and into the Manager. Here, the distinction is not incredibly
important, but it will become important as we make more operations
asynchronous, and need to be aware of which services may be in the
process of, say, shutting down for a restart so that we do not try to
start them again too soon.

This "reconciliation" logic was also extracted into a pure function
for ease of testability.

We also introduce a new `SpecDir` type to model the spec directory
itself, as well as operations on it and the files within. Previously,
this functionality was split between the `SpecWatcher` and the
`Manager`. In particular, this obscured some of the core logic carried
out by the `SpecWatcher`.

The overall structure of the `SpecWatcher` was also simplified, and
removes an extra thread; there were three, now there are two. Two are
from the underlying `notify` crate, while the extra one was one of our
own creation that turned out to not be strictly necessary. Rather than
keeping a thread around to poll the notifier thread and store the
result in an atomic boolean, now we just keep ownership of the other
side of the notification channel and ask it if it has any events.

Signed-off-by: Christopher Maier <cmaier@chef.io>